+++ src/selkies_gstreamer/__main__.py
@@ -10,9 +10,10 @@ logger_gstwebrtc_app_resize = logging.getLogger("gstwebrtc_app_resize")
 logger_signaling = logging.getLogger("signaling")
 logger_system_monitor = logging.getLogger("system_monitor")
 logger_webrtc_input = logging.getLogger("webrtc_input")
-logger_webrtc_signalling = logging.getLogger("signalling")
+logger_webrtc_signalling = logging.getLogger("webrtc_signalling")
 logger = logging.getLogger("main")
 web_logger = logging.getLogger("web")
+data_logger = logging.getLogger("data_websocket")
 
 # Imports
 import concurrent.futures
@@ -49,9 +50,6 @@ from shutil import which
 from signal import SIGINT, signal
 from watchdog.events import FileClosedEvent, FileSystemEventHandler
 from watchdog.observers import Observer
-from Xlib import display
-from Xlib import X
-from Xlib.ext import xfixes, xtest
 from prometheus_client import Gauge, Histogram, Info, start_http_server
 try:
     import gi
@@ -69,10 +67,10 @@ If GStreamer is installed at a certain location, set the path to the environment
 export GSTREAMER_PATH="${GSTREAMER_PATH:-$(pwd)}"
 export PATH="${GSTREAMER_PATH}/bin${PATH:+:${PATH}}"
 export LD_LIBRARY_PATH="${GSTREAMER_PATH}/lib/x86_64-linux-gnu${LD_LIBRARY_PATH:+:${LD_LIBRARY_PATH}}"
-export GST_PLUGIN_PATH="${GSTREAMER_PATH}/lib/x86_64-linux-gnu/gstreamer-1.0${GST_PLUGIN_PATH:+:${GST_PLUGIN_PATH}}"
-export GST_PLUGIN_SYSTEM_PATH="${XDG_DATA_HOME:-${HOME:-~}/.local/share}/gstreamer-1.0/plugins:/usr/lib/x86_64-linux-gnu/gstreamer-1.0${GST_PLUGIN_SYSTEM_PATH:+:${GST_PLUGIN_SYSTEM_PATH}}"
-export GI_TYPELIB_PATH="${GSTREAMER_PATH}/lib/x86_64-linux-gnu/girepository-1.0:/usr/lib/x86_64-linux-gnu/girepository-1.0${GI_TYPELIB_PATH:+:${GI_TYPELIB_PATH}}"
-export PYTHONPATH="${GSTREAMER_PATH}/lib/python3/dist-packages${PYTHONPATH:+:${PYTHONPATH}}"
+export GST_PLUGIN_PATH="${GSTREAMER_PATH}/lib/x86_64-linux-gnu/gstreamer-1.0${GST_PLUGIN_PATH:+:${PATH}}"
+export GST_PLUGIN_SYSTEM_PATH="${XDG_DATA_HOME:-${HOME:-~}/.local/share}/gstreamer-1.0/plugins:/usr/lib/x86_64-linux-gnu/gstreamer-1.0${GST_PLUGIN_SYSTEM_PATH:+:${PATH}}"
+export GI_TYPELIB_PATH="${GSTREAMER_PATH}/lib/x86_64-linux-gnu/girepository-1.0:/usr/lib/x86_64-linux-gnu/girepository-1.0${GI_TYPELIB_PATH:+:${PATH}}"
+export PYTHONPATH="${GSTREAMER_PATH}/lib/python3/dist-packages${PYTHONPATH:+:${PATH}}"
 Replace "x86_64-linux-gnu" in other architectures manually or use "$(gcc -print-multiarch)" in place.
 """
     logger_gstwebrtc_app.error(msg)
@@ -80,116 +78,13 @@ Replace "x86_64-linux-gnu" in other architectures manually or use "$(gcc -print-
     sys.exit(1)
 logger_gstwebrtc_app.info("GStreamer-Python install looks OK")
 import GPUtil
-import pynput
-from PIL import Image
 
-# Constants
-EV_SYN = 0x00
-EV_KEY = 0x01
-EV_REL = 0x02
-EV_ABS = 0x03
-EV_MSC = 0x04
-EV_SW = 0x05
-EV_LED = 0x11
-EV_SND = 0x12
-EV_REP = 0x14
-EV_FF = 0x15
-EV_PWR = 0x16
-EV_FF_STATUS = 0x17
-EV_MAX = 0x1F
-EV_CNT = EV_MAX + 1
-SYN_REPORT = 0
-SYN_CONFIG = 1
-SYN_MT_REPORT = 2
-SYN_DROPPED = 3
-SYN_MAX = 0xF
-SYN_CNT = SYN_MAX + 1
-BTN_MISC = 0x100
-BTN_0 = 0x100
-BTN_1 = 0x101
-BTN_2 = 0x102
-BTN_3 = 0x103
-BTN_4 = 0x104
-BTN_5 = 0x105
-BTN_6 = 0x106
-BTN_7 = 0x107
-BTN_8 = 0x108
-BTN_9 = 0x109
-BTN_MOUSE = 0x110
-BTN_LEFT = 0x110
-BTN_RIGHT = 0x111
-BTN_MIDDLE = 0x112
-BTN_SIDE = 0x113
-BTN_EXTRA = 0x114
-BTN_FORWARD = 0x115
-BTN_BACK = 0x116
-BTN_TASK = 0x117
-BTN_JOYSTICK = 0x120
-BTN_TRIGGER = 0x120
-BTN_THUMB = 0x121
-BTN_THUMB2 = 0x122
-BTN_TOP = 0x123
-BTN_TOP2 = 0x124
-BTN_PINKIE = 0x125
-BTN_BASE = 0x126
-BTN_BASE2 = 0x127
-BTN_BASE3 = 0x128
-BTN_BASE4 = 0x129
-BTN_BASE5 = 0x12A
-BTN_BASE6 = 0x12B
-BTN_DEAD = 0x12F
-BTN_GAMEPAD = 0x130
-BTN_SOUTH = 0x130
-BTN_A = BTN_SOUTH
-BTN_EAST = 0x131
-BTN_B = BTN_EAST
-BTN_C = 0x132
-BTN_NORTH = 0x133
-BTN_X = BTN_NORTH
-BTN_WEST = 0x134
-BTN_Y = BTN_WEST
-BTN_Z = 0x135
-BTN_TL = 0x136
-BTN_TR = 0x137
-BTN_TL2 = 0x138
-BTN_TR2 = 0x139
-BTN_SELECT = 0x13A
-BTN_START = 0x13B
-BTN_MODE = 0x13C
-BTN_THUMBL = 0x13D
-BTN_THUMBR = 0x13E
-ABS_X = 0x00
-ABS_Y = 0x01
-ABS_Z = 0x02
-ABS_RX = 0x03
-ABS_RY = 0x04
-ABS_RZ = 0x05
-ABS_THROTTLE = 0x06
-ABS_RUDDER = 0x07
-ABS_WHEEL = 0x08
-ABS_GAS = 0x09
-ABS_BRAKE = 0x0A
-ABS_HAT0X = 0x10
-ABS_HAT0Y = 0x11
-ABS_HAT1X = 0x12
-ABS_HAT1Y = 0x13
-ABS_HAT2X = 0x14
-ABS_HAT2Y = 0x15
-ABS_HAT3X = 0x16
-ABS_HAT3Y = 0x17
-ABS_PRESSURE = 0x18
-ABS_DISTANCE = 0x19
-ABS_TILT_X = 0x1A
-ABS_TILT_Y = 0x1B
-ABS_TOOL_WIDTH = 0x1C
-ABS_VOLUME = 0x20
-ABS_PROFILE = 0x21
-JS_EVENT_BUTTON = 0x01
-JS_EVENT_AXIS = 0x02
-MAX_BTNS = 512
-MAX_AXES = 64
-ABS_MIN = -32767
-ABS_MAX = 32767
+
+# Import the new input_handler module
+from input_handler import WebRTCInput, SelkiesGamepad, GamepadMapper, get_btn_event, get_axis_event, detect_gamepad_config, get_num_btns_for_mapping, get_num_axes_for_mapping, normalize_axis_val, normalize_trigger_val, ABS_MIN, ABS_MAX, STANDARD_XPAD_CONFIG, XPAD_CONFIG_MAP, logger_selkies_gamepad
+
+
+# Constants (Some constants are moved to input_handler.py, keep only those not moved if necessary for main)
 FPS_HIST_BUCKETS = (0, 20, 40, 60)
 DEFAULT_RTC_CONFIG = """{
   "lifetimeDuration": "86400s",
@@ -209,62 +104,10 @@ MIME_TYPES = {
     "css": "text/css",
     "ico": "image/x-icon",
 }
-MOUSE_POSITION = 10
-MOUSE_MOVE = 11
-MOUSE_SCROLL_UP = 20
-MOUSE_SCROLL_DOWN = 21
-MOUSE_BUTTON_PRESS = 30
-MOUSE_BUTTON_RELEASE = 31
-MOUSE_BUTTON = 40
-MOUSE_BUTTON_LEFT = 41
-MOUSE_BUTTON_MIDDLE = 42
-MOUSE_BUTTON_RIGHT = 43
-UINPUT_BTN_LEFT = (0x01, 0x110)
-UINPUT_BTN_MIDDLE = (0x01, 0x112)
-UINPUT_BTN_RIGHT = (0x01, 0x111)
-UINPUT_REL_X = (0x02, 0x00)
-UINPUT_REL_Y = (0x02, 0x01)
-UINPUT_REL_WHEEL = (0x02, 0x08)
-MOUSE_BUTTON_MAP = {
-    MOUSE_BUTTON_LEFT: {
-        "uinput": UINPUT_BTN_LEFT,
-        "pynput": pynput.mouse.Button.left,
-    },
-    MOUSE_BUTTON_MIDDLE: {
-        "uinput": UINPUT_BTN_MIDDLE,
-        "pynput": pynput.mouse.Button.middle,
-    },
-    MOUSE_BUTTON_RIGHT: {
-        "uinput": UINPUT_BTN_RIGHT,
-        "pynput": pynput.mouse.Button.right,
-    },
-}
 
-def get_btn_event(btn_num, btn_val):
-    ts = int((time.time() * 1000) % 1000000000)
-    struct_format = "IhBB"
-    event = struct.pack(struct_format, ts, btn_val, JS_EVENT_BUTTON, btn_num)
-    logger_selkies_gamepad.debug(struct.unpack(struct_format, event))
-    return event
-def get_axis_event(axis_num, axis_val):
-    ts = int((time.time() * 1000) % 1000000000)
-    struct_format = "IhBB"
-    event = struct.pack(struct_format, ts, axis_val, JS_EVENT_AXIS, axis_num)
-    logger_selkies_gamepad.debug(struct.unpack(struct_format, event))
-    return event
-def detect_gamepad_config(name):
-    return STANDARD_XPAD_CONFIG
-def get_num_btns_for_mapping(cfg):
-    num_mapped_btns = len(
-        [i for j in cfg["mapping"]["axes_to_btn"].values() for i in j]
-    )
-    return len(cfg["btn_map"]) + num_mapped_btns
-def get_num_axes_for_mapping(cfg):
-    return len(cfg["axes_map"])
-def normalize_axis_val(val):
-    return round(ABS_MIN + ((val + 1) * (ABS_MAX - ABS_MIN)) / 2)
-def normalize_trigger_val(val):
-    return round(val * (ABS_MAX - ABS_MIN)) + ABS_MIN
+
+# ... (Rest of the code, removing input related functions and classes that are now in input_handler.py) ...
+
 def fit_res(w, h, max_w, max_h):
     if w < max_w and h < max_h:
         return w, h
@@ -725,203 +568,7 @@ def set_json_app_argument(config_path, key, value):
         json.dump(json_data, f)
     return True
 
-STANDARD_XPAD_CONFIG = {
-    "name": "Selkies Controller",
-    "btn_map": [
-        BTN_A,
-        BTN_B,
-        BTN_X,
-        BTN_Y,
-        BTN_TL,
-        BTN_TR,
-        BTN_SELECT,
-        BTN_START,
-        BTN_MODE,
-        BTN_THUMBL,
-        BTN_THUMBR,
-    ],
-    "axes_map": [ABS_X, ABS_Y, ABS_Z, ABS_RX, ABS_RY, ABS_RZ, ABS_HAT0X, ABS_HAT0Y],
-    "mapping": {
-        "axes_to_btn": {2: (6,), 5: (7,), 6: (15, 14), 7: (13, 12)},
-        "axes": {
-            2: 3,
-            3: 4,
-        },
-        "btns": {8: 6, 9: 7, 10: 9, 11: 10, 16: 8},
-        "trigger_axes": [2, 5],
-    },
-}
-XPAD_CONFIG_MAP = {
-    ("045e", "0b12"): STANDARD_XPAD_CONFIG,
-}
-class SelkiesGamepad:
-    def __init__(self, socket_path):
-        self.socket_path = socket_path
-        self.mapper = None
-        self.name = None
-        self.server = None
-        self.config = None
-        self.clients = {}
-        self.events = asyncio.Queue() # Changed to asyncio.Queue
-        self.running = False
-    def set_config(self, name, num_btns, num_axes):
-        self.name = name
-        self.config = detect_gamepad_config(name)
-        self.mapper = GamepadMapper(self.config, name, num_btns, num_axes)
-    def __make_config(self):
-        if not self.config:
-            logger_selkies_gamepad.error(
-                "could not make js config because it has not yet been set."
-            )
-            return None
-        num_btns = len(self.config["btn_map"])
-        num_axes = len(self.config["axes_map"])
-        btn_map = [i for i in self.config["btn_map"]]
-        axes_map = [i for i in self.config["axes_map"]]
-        btn_map[num_btns:MAX_BTNS] = [0 for i in range(num_btns, MAX_BTNS)]
-        axes_map[num_axes:MAX_AXES] = [0 for i in range(num_axes, MAX_AXES)]
-        struct_fmt = "255sHH%dH%dB" % (MAX_BTNS, MAX_AXES)
-        data = struct.pack(
-            struct_fmt,
-            self.config["name"].encode(),
-            num_btns,
-            num_axes,
-            *btn_map,
-            *axes_map,
-        )
-        return data
-    async def __send_events(self):
-        while self.running:
-            if self.events.empty():
-                await asyncio.sleep(0.01) # Increased sleep duration to 0.01 for less aggressive polling.
-                continue
-            while self.running and not self.events.empty():
-                event = await self.events.get() # Await directly from asyncio.Queue
-                await self.send_event(event)
-                self.events.task_done() # Indicate task completion for Queue
-    def send_btn(self, btn_num, btn_val):
-        if not self.mapper:
-            logger_selkies_gamepad.warning(
-                "failed to send js button event because mapper was not set"
-            )
-            return
-        event = self.mapper.get_mapped_btn(btn_num, btn_val)
-        if event is not None:
-            self.events.put_nowait(event) # Use put_nowait for non-blocking queue operation
-    def send_axis(self, axis_num, axis_val):
-        if not self.mapper:
-            logger_selkies_gamepad.warning(
-                "failed to send js axis event because mapper was not set"
-            )
-            return
-        event = self.mapper.get_mapped_axis(axis_num, axis_val)
-        if event is not None:
-            self.events.put_nowait(event) # Use put_nowait for non-blocking queue operation
-    async def send_event(self, event):
-        if len(self.clients) < 1:
-            return
-        closed_clients = []
-        for fd in self.clients:
-            try:
-                client = self.clients[fd]
-                logger_selkies_gamepad.debug("Sending event to client with fd: %d" % fd)
-                # Directly await client.send instead of using asyncio.to_thread, assuming client is asyncio socket
-                await asyncio.get_running_loop().sock_sendall(client, event) # Use sock_sendall for asyncio socket
-            except BrokenPipeError:
-                logger_selkies_gamepad.info("Client %d disconnected" % fd)
-                closed_clients.append(fd)
-                client.close()
-        for fd in closed_clients:
-            del self.clients[fd]
-    async def setup_client(self, client):
-        logger_selkies_gamepad.info(
-            "Sending config to client with fd: %d" % client.fileno()
-        )
-        try:
-            config_data = self.__make_config()
-            if not config_data:
-                return
-            # Directly await client.send instead of using asyncio.to_thread, assuming client is asyncio socket
-            await asyncio.get_running_loop().sock_sendall(client, config_data) # Use sock_sendall for asyncio socket
-            await asyncio.sleep(0.5)
-            for btn_num in range(len(self.config["btn_map"])):
-                self.send_btn(btn_num, 0)
-            for axis_num in range(len(self.config["axes_map"])):
-                self.send_axis(axis_num, 0)
-        except BrokenPipeError:
-            client.close()
-            logger_selkies_gamepad.info("Client disconnected")
-    async def run_server(self):
-        try:
-            os.unlink(self.socket_path)
-        except OSError:
-            if os.path.exists(self.socket_path):
-                raise
-        self.server = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
-        self.server.bind(self.socket_path)
-        self.server.listen(1)
-        self.server.setblocking(False)
-        logger_selkies_gamepad.info(
-            "Listening for connections on %s" % self.socket_path
-        )
-        asyncio.create_task(self.__send_events())
-        self.running = True
-        try:
-            while self.running:
-                try:
-                    client, _ = await asyncio.get_running_loop().sock_accept(self.server) # Use sock_accept for asyncio socket
-                except asyncio.TimeoutError:
-                    continue
-                fd = client.fileno()
-                logger_selkies_gamepad.info("Client connected with fd: %d" % fd)
-                await self.setup_client(client)
-                self.clients[fd] = client
-        finally:
-            self.server.close()
-            try:
-                os.unlink(self.socket_path)
-            except:
-                pass
-        logger_selkies_gamepad.info(
-            "Stopped gamepad socket server for %s" % self.socket_path
-        )
-class GamepadMapper:
-    def __init__(self, config, name, num_btns, num_axes):
-        self.config = config
-        self.input_name = name
-        self.input_num_btns = num_btns
-        self.input_num_axes = num_axes
-    def get_mapped_btn(self, btn_num, btn_val):
-        axis_num = None
-        axis_sign = 1
-        for axis, mapping in self.config["mapping"]["axes_to_btn"].items():
-            if btn_num in mapping:
-                axis_num = axis
-                if len(mapping) > 1:
-                    axis_sign = 1 if mapping[0] == btn_num else -1
-                break
-        if axis_num is not None:
-            axis_val = normalize_axis_val(btn_val * axis_sign)
-            if axis_num in self.config["mapping"]["trigger_axes"]:
-                axis_val = normalize_trigger_val(btn_val)
-            return get_axis_event(axis_num, axis_val)
-        mapped_btn = self.config["mapping"]["btns"].get(btn_num, btn_num)
-        if mapped_btn >= len(self.config["btn_map"]):
-            logger_selkies_gamepad.error(
-                "cannot send button num %d, max num buttons is %d"
-                % (mapped_btn, len(self.config["btn_map"]) - 1)
-            )
-            return None
-        return get_btn_event(mapped_btn, int(btn_val))
-    def get_mapped_axis(self, axis_num, axis_val):
-        mapped_axis = self.config["mapping"]["axes"].get(axis_num, axis_num)
-        if mapped_axis >= len(self.config["axes_map"]):
-            logger_selkies_gamepad.error(
-                "cannot send axis %d, max axis num is %d"
-                % (mapped_axis, len(self.config["axes_map"]) - 1)
-            )
-            return None
-        return get_axis_event(mapped_axis, normalize_axis_val(axis_val))
+
 class GPUMonitor:
     def __init__(self, period=1, enabled=True):
         self.period = period
@@ -942,6 +589,46 @@ class GPUMonitor:
         logger_gpu_monitor.info("GPU monitor stopped")
     def stop(self):
         self.running = False
+
+class DataStreamingServer: # New DataStreamingServer class
+    def __init__(self, port):
+        self.port = port
+        self.server = None
+        self.stop_server = None
+        self.data_ws = None
+
+    async def ws_handler(self, websocket):
+        raddr = websocket.remote_address
+        data_logger.info(f"Data WebSocket connected from {raddr}")
+        self.data_ws = websocket # Store the websocket connection
+
+        try:
+            async for message in websocket:
+                # We expect to only send data, not receive messages.
+                pass # Keep connection alive
+        except websockets.exceptions.ConnectionClosed:
+            data_logger.info(f"Data WebSocket disconnected from {raddr}")
+        finally:
+            self.data_ws = None # Clear connection on disconnect
+
+    async def run_server(self):
+        self.stop_server = asyncio.Future()
+        async with websockets.asyncio.server.serve(
+            self.ws_handler, '0.0.0.0', self.port, compression=None
+        ) as self.server:
+            data_logger.info(f"Data WebSocket Server listening on port {self.port}")
+            await self.stop_server
+
+    async def stop(self):
+        data_logger.info("Stopping Data WebSocket Server...")
+        if self.stop_server is not None:
+            self.stop_server.set_result(True)
+        if self.server:
+            self.server.close()
+            await self.server.wait_closed()
+        data_logger.info("Data WebSocket Server Stopped.")
+
+
 class GSTWebRTCAppError(Exception):
     pass
 class GSTWebRTCApp:
@@ -960,6 +647,7 @@ class GSTWebRTCApp:
         congestion_control=False,
         video_packetloss_percent=0.0,
         audio_packetloss_percent=0.0,
+        data_streaming_server=None
     ):
         self.async_event_loop = async_event_loop
         self.stun_servers = stun_servers
@@ -1020,6 +708,77 @@ class GSTWebRTCApp:
         self.ximagesrc = None
         self.ximagesrc_caps = None
         self.last_cursor_sent = None
+        self.data_streaming_server = data_streaming_server
+
+    def ws_pipeline(self, keyframe_distance, keyframe_frame_distance, framerate, vbv_multiplier_sw, fec_video_bitrate):
+        pipeline_string = f"""
+            ximagesrc name=source ! queue name=queue1 ! videoconvert name=convert !
+            capsfilter name=videoconvert_capsfilter caps=video/x-raw,format=NV12 !
+            x264enc name=encoder
+                threads={min(4, max(1, len(os.sched_getaffinity(0)) - 1))}
+                aud=false b-adapt=false bframes=0 dct8x8=false insert-vui=true
+                key-int-max={2147483647 if keyframe_distance == -1.0 else keyframe_frame_distance}
+                mb-tree=false rc-lookahead=0 sync-lookahead=0
+                vbv-buf-capacity={int((1000 + framerate - 1) // framerate * vbv_multiplier_sw)}
+                sliced-threads=true byte-stream=true pass=cbr speed-preset=ultrafast tune=zerolatency
+                bitrate={fec_video_bitrate} !
+            queue name=queue2 ! appsink name=sink emit-signals=true sync=false
+        """
+
+        try:
+            pipeline = Gst.parse_launch(pipeline_string)
+        except Gst.ParseError as e:
+            print(f"Error parsing pipeline string: {e}")
+            return # Exit function on error
+
+        if not pipeline:
+            print("Error: Could not create pipeline from string")
+            return # Exit function on error
+
+        source = pipeline.get_by_name("source")
+        if not source:
+            print("Error: Could not get source element")
+            return # Exit function on error
+        source.set_property("show-pointer", 0)
+        source.set_property("remote", 1)
+
+        sink = pipeline.get_by_name("sink")
+        if not sink:
+            print("Error: Could not get sink element")
+            return # Exit function on error
+        def on_new_sample(sink):
+            sample = sink.emit("pull-sample")
+            if sample:
+                buffer = sample.get_buffer()
+                if buffer:
+                    caps = sample.get_caps()
+                    if caps:
+                        structure = caps.get_structure(0)
+                        codec_name = structure.get_name()
+                    else:
+                        codec_name = "Unknown Codec"
+                    is_delta_frame = bool(buffer.get_flags() & Gst.BufferFlags.DELTA_UNIT)
+                    frame_type = "Deltaframe" if is_delta_frame else "Keyframe"
+                    success, map_info = buffer.map(Gst.MapFlags.READ)
+                    if success:
+                        data = map_info.data
+                        data_copy = bytes(data)
+                        frame_type_byte = b'\x00' if is_delta_frame else b'\x01'
+                        prefixed_data = frame_type_byte + data_copy
+                        if self.data_streaming_server.data_ws:
+                            try:
+                                asyncio.run_coroutine_threadsafe(self.data_streaming_server.data_ws.send(prefixed_data), self.async_event_loop)
+                                data_logger.debug(f"Sent {len(prefixed_data)} bytes (including flag) to data websocket")
+                            except Exception as e:
+                                data_logger.error(f"Error sending data over websocket: {e}")
+                        buffer.unmap(map_info)
+                    else:
+                        print("Error mapping buffer")
+                return Gst.FlowReturn.OK
+            return Gst.FlowReturn.ERROR
+        sink.connect("new-sample", on_new_sample)
+        pipeline.set_state(Gst.State.PLAYING)
+        print("Pipeline is PLAYING... (ws_pipeline function version - check console for byte and metadata output)")
     def stop_ximagesrc(self):
         if self.ximagesrc:
             self.ximagesrc.set_state(Gst.State.NULL)
@@ -1293,7 +1052,6 @@ class GSTWebRTCApp:
                     * self.vbv_multiplier_va
                 ),
             )
-            vah264enc.set_property("dct8x8", False)
             vah264enc.set_propertyLOGLEVEL(
                 "key-int-max",
                 1024
@@ -1454,6 +1212,18 @@ class GSTWebRTCApp:
             videoconvert_caps.set_value("format", "NV12")
             videoconvert_capsfilter = Gst.ElementFactory.make("capsfilter")
             videoconvert_capsfilter.set_property("caps", videoconvert_caps)
+
+            async def run_ws_pipeline_async(): # Define an async wrapper function inside the method
+                await asyncio.to_thread(
+                    self.ws_pipeline,
+                    self.keyframe_distance,
+                    self.keyframe_frame_distance,
+                    self.framerate,
+                    self.vbv_multiplier_sw,
+                    self.fec_video_bitrate
+                )
+            asyncio.create_task(run_ws_pipeline_async()) # Run the async wrapper synchronously
+
             x264enc = Gst.ElementFactory.make("x264enc", "x264enc")
             x264enc.set_property(
                 "threads", min(4, max(1, len(os.sched_getaffinity(0)) - 1))
@@ -1558,7 +1328,7 @@ class GSTWebRTCApp:
                 vpenc = Gst.ElementFactory.make("vp9enc", "vpenc")
                 vpenc.set_property("frame-parallel-decoding", True)
                 vpenc.set_property("row-mt", True)
-            vpenc.set_property(
+            vpenc.set_property(keyframe_distance,
                 "threads", min(16, max(1, len(os.sched_getaffinity(0)) - 1))
             )
             vbv_buffer_size = int(
@@ -2068,6 +1838,7 @@ class GSTWebRTCApp:
         promise = Gst.Promise.new()
         self.webrtcbin.emit("set-remote-description", answer, promise)
         promise.interrupt()
+        pass # sdp_text = offer.sdp.as_text() # No need to process offer SDP on answer
     def set_ice(self, mlineindex, candidate):
         logger_gstwebrtc_app.info(
             "setting ICE candidate: %d, %s" % (mlineindex, candidate)
@@ -2852,8 +2623,9 @@ class Metrics:
         )
         self.prev_stats_video_header_len = None
         self.prev_stats_audio_header_len = None
+
 class WebRTCSimpleServer(object):
-    def __init__(self, options):
+    def __init__(self, options): # Pass DataStreamingServer instance
         self.peers = dict()
         self.sessions = dict()
         self.rooms = dict()
@@ -2957,6 +2729,8 @@ class WebRTCSimpleServer(object):
                     response_headers,
                     b"Authorization required",
                 )
+        if path == "/websocket": # Handle new data websocket path
+            return None # No HTTP response for websocket upgrade
         if (
             path == "/ws/"
             or path == "/ws"
@@ -3312,588 +3086,6 @@ class SystemMonitor:
         logger_system_monitor.info("system monitor stopped")
     def stop(self):
         self.running = False
-class WebRTCInputError(Exception):
-    pass
-class WebRTCInput:
-    def __init__(
-        self,
-        uinput_mouse_socket_path="",
-        js_socket_path="",
-        enable_clipboard="",
-        enable_cursors=True,
-        cursor_size=16,
-        cursor_scale=1.0,
-        cursor_debug=False,
-    ):
-        self.clipboard_running = False
-        self.uinput_mouse_socket_path = uinput_mouse_socket_path
-        self.uinput_mouse_socket = None
-        self.js_socket_path_map = {
-            i: os.path.join(js_socket_path, "selkies_js%d.sock" % i) for i in range(4)
-        }
-        self.js_map = {}
-        self.enable_clipboard = enable_clipboard
-        self.enable_cursors = enable_cursors
-        self.cursors_running = False
-        self.cursor_cache = {}
-        self.cursor_scale = cursor_scale
-        self.cursor_size = cursor_size
-        self.cursor_debug = cursor_debug
-        self.keyboard = None
-        self.mouse = None
-        self.joystick = None
-        self.xdisplay = None
-        self.button_mask = 0
-        self.ping_start = None
-        self.on_video_encoder_bit_rate = lambda bitrate: logger_webrtc_input.warning(
-            "unhandled on_video_encoder_bit_rate"
-        )
-        self.on_audio_encoder_bit_rate = lambda bitrate: logger_webrtc_input.warning(
-            "unhandled on_audio_encoder_bit_rate"
-        )
-        self.on_mouse_pointer_visible = lambda visible: logger_webrtc_input.warning(
-            "unhandled on_mouse_pointer_visible"
-        )
-        self.on_clipboard_read = lambda data: logger_webrtc_input.warning(
-            "unhandled on_clipboard_read"
-        )
-        self.on_set_fps = lambda fps: logger_webrtc_input.warning(
-            "unhandled on_set_fps"
-        )
-        self.on_set_enable_resize = (
-            lambda enable_resize, res: logger_webrtc_input.warning(
-                "unhandled on_set_enable_resize"
-            )
-        )
-        self.on_client_fps = lambda fps: logger_webrtc_input.warning(
-            "unhandled on_client_fps"
-        )
-        self.on_client_latency = lambda latency: logger_webrtc_input.warning(
-            "unhandled on_client_latency"
-        )
-        self.on_resize = lambda res: logger_webrtc_input.warning("unhandled on_resize")
-        self.on_scaling_ratio = lambda res: logger_webrtc_input.warning(
-            "unhandled on_scaling_ratio"
-        )
-        self.on_ping_response = lambda latency: logger_webrtc_input.warning(
-            "unhandled on_ping_response"
-        )
-        self.on_cursor_change = lambda msg: logger_webrtc_input.warning(
-            "unhandled on_cursor_change"
-        )
-        self.on_client_webrtc_stats = (
-            lambda webrtc_stat_type, webrtc_stats: logger_webrtc_input.warning(
-                "unhandled on_client_webrtc_stats"
-            )
-        )
-    def __keyboard_connect(self):
-        self.keyboard = pynput.keyboard.Controller()
-    def __mouse_connect(self):
-        if self.uinput_mouse_socket_path:
-            logger_webrtc_input.info(
-                "Connecting to uinput mouse socket: %s" % self.uinput_mouse_socket_path
-            )
-            self.uinput_mouse_socket = socket.socket(socket.AF_UNIX, socket.SOCK_DGRAM)
-        self.mouse = pynput.mouse.Controller()
-    def __mouse_disconnect(self):
-        if self.mouse:
-            del self.mouse
-            self.mouse = None
-    def __mouse_emit(self, *args, **kwargs):
-        if self.uinput_mouse_socket_path:
-            cmd = {"args": args, "kwargs": kwargs}
-            data = msgpack.packb(cmd, use_bin_type=True)
-            self.uinput_mouse_socket.sendto(data, self.uinput_mouse_socket_path)
-    def __js_connect(self, js_num, name, num_btns, num_axes):
-        logger_webrtc_input.info(
-            "creating selkies gamepad for js%d, name: '%s', buttons: %d, axes: %d"
-            % (js_num, name, num_btns, num_axes)
-        )
-        socket_path = self.js_socket_path_map.get(js_num, None)
-        if socket_path is None:
-            logger_webrtc_input.error(
-                "failed to connect js%d because socket_path was not found" % js_num
-            )
-            return
-        js = SelkiesGamepad(socket_path)
-        js.set_config(name, num_btns, num_axes)
-        asyncio.create_task(js.run_server())
-        self.js_map[js_num] = js
-    async def __js_disconnect(self, js_num=None):
-        if js_num is None:
-            for js in self.js_map.values():
-                await asyncio.to_thread(js.stop_server)
-            self.js_map = {}
-            return
-        js = await asyncio.to_thread(self.js_map.get, js_num, None)
-        if js is not None:
-            logger_webrtc_input.info("stopping gamepad %d" % js_num)
-            await asyncio.to_thread(js.stop_server)
-            del self.js_map[js_num]
-    def __js_emit_btn(self, js_num, btn_num, btn_val):
-        js = self.js_map.get(js_num, None)
-        if js is None:
-            logger_webrtc_input.error(
-                "cannot send button because js%d is not connected" % js_num
-            )
-            return
-        logger_webrtc_input.debug(
-            "sending js%d button num %d with val %d" % (js_num, btn_num, btn_val)
-        )
-        js.send_btn(btn_num, btn_val)
-    def __js_emit_axis(self, js_num, axis_num, axis_val):
-        js = self.js_map.get(js_num, None)
-        if js is None:
-            logger_webrtc_input.error(
-                "cannot send axis because js%d is not connected" % js_num
-            )
-            return
-        logger_webrtc_input.debug(
-            "sending js%d axis num %d with val %d" % (js_num, axis_num, axis_val)
-        )
-        js.send_axis(axis_num, axis_val)
-    async def connect(self):
-        self.xdisplay = display.Display()
-        self.__keyboard_connect()
-        self.reset_keyboard()
-        self.__mouse_connect()
-    async def disconnect(self):
-        await self.__js_disconnect()
-        self.__mouse_disconnect()
-    def reset_keyboard(self):
-        logger_webrtc_input.info("Resetting keyboard modifiers.")
-        lctrl = 65507
-        lshift = 65505
-        lalt = 65513
-        rctrl = 65508
-        rshift = 65506
-        ralt = 65027
-        lmeta = 65511
-        rmeta = 65512
-        keyf = 102
-        keyF = 70
-        keym = 109
-        keyM = 77
-        escape = 65307
-        for k in [
-            lctrl,
-            lshift,
-            lalt,
-            rctrl,
-            rshift,
-            ralt,
-            lmeta,
-            rmeta,
-            keyf,
-            keyF,
-            keym,
-            keyM,
-            escape,
-        ]:
-            self.send_x11_keypress(k, down=False)
-    def send_mouse(self, action, data):
-        if action == MOUSE_POSITION:
-            if self.mouse:
-                self.mouse.position = data
-        elif action == MOUSE_MOVE:
-            x, y = data
-            if self.uinput_mouse_socket_path:
-                self.__mouse_emit(UINPUT_REL_X, x, syn=False)
-                self.__mouse_emit(UINPUT_REL_Y, y)
-            else:
-                xtest.fake_input(
-                    self.xdisplay,
-                    Xlib.X.MotionNotify,
-                    detail=True,
-                    root=Xlib.X.NONE,
-                    x=x,
-                    y=y,
-                )
-                self.xdisplay.sync()
-        elif action == MOUSE_SCROLL_UP:
-            if self.uinput_mouse_socket_path:
-                self.__mouse_emit(UINPUT_REL_WHEEL, 1)
-            else:
-                self.mouse.scroll(0, -1)
-        elif action == MOUSE_SCROLL_DOWN:
-            if self.uinput_mouse_socket_path:
-                self.__mouse_emit(UINPUT_REL_WHEEL, -1)
-            else:
-                self.mouse.scroll(0, 1)
-        elif action == MOUSE_BUTTON:
-            if self.uinput_mouse_socket_path:
-                btn = MOUSE_BUTTON_MAP[data[1]]["uinput"]
-            else:
-                btn = MOUSE_BUTTON_MAP[data[1]]["pynput"]
-            if data[0] == MOUSE_BUTTON_PRESS:
-                if self.uinput_mouse_socket_path:
-                    self.__mouse_emit(btn, 1)
-                else:
-                    self.mouse.press(btn)
-            else:
-                if self.uinput_mouse_socket_path:
-                    self.__mouse_emit(btn, 0)
-                else:
-                    self.mouse.release(btn)
-    def send_x11_keypress(self, keysym, down=True):
-        try:
-            if keysym == 60 and self.keyboard._display.keysym_to_keycode(keysym) == 94:
-                keysym = 44
-            keycode = pynput.keyboard.KeyCode(keysym)
-            if down:
-                self.keyboard.press(keycode)
-            else:
-                self.keyboard.release(keycode)
-        except Exception as e:
-            logger_webrtc_input.error("failed to send keypress: {}".format(e))
-    def send_x11_mouse(self, x, y, button_mask, scroll_magnitude, relative=False):
-        if relative:
-            self.send_mouse(MOUSE_MOVE, (x, y))
-        else:
-            self.send_mouse(MOUSE_POSITION, (x, y))
-        if button_mask != self.button_mask:
-            max_buttons = 5
-            for i in range(0, max_buttons):
-                if (button_mask ^ self.button_mask) & (1 << i):
-                    action = MOUSE_BUTTON
-                    btn_action = MOUSE_BUTTON_PRESS
-                    btn_num = MOUSE_BUTTON_LEFT
-                    if button_mask & (1 << i):
-                        btn_action = MOUSE_BUTTON_PRESS
-                    else:
-                        btn_action = MOUSE_BUTTON_RELEASE
-                    if i == 1:
-                        btn_num = MOUSE_BUTTON_MIDDLE
-                    elif i == 2:
-                        btn_num = MOUSE_BUTTON_RIGHT
-                    elif i == 3 and button_mask != 0:
-                        action = MOUSE_SCROLL_UP
-                    elif i == 4 and button_mask != 0:
-                        action = MOUSE_SCROLL_DOWN
-                    data = (btn_action, btn_num)
-                    if i == 3 or i == 4:
-                        for i in range(1, scroll_magnitude):
-                            self.send_mouse(action, data)
-                    self.send_mouse(action, data)
-            self.button_mask = button_mask
-        if not relative:
-            self.xdisplay.sync()
-    def read_clipboard(self):
-        try:
-            result = subprocess.run(
-                ("xsel", "--clipboard", "--output"),
-                check=True,
-                text=True,
-                capture_output=True,
-                timeout=3,
-            )
-            return result.stdout
-        except subprocess.SubprocessError as e:
-            logger_webrtc_input.warning(f"Error while capturing clipboard: {e}")
-    def write_clipboard(self, data):
-        try:
-            subprocess.run(
-                ("xsel", "--clipboard", "--input"),
-                input=data.encode(),
-                check=True,
-                timeout=3,
-            )
-            return True
-        except subprocess.SubprocessError as e:
-            logger_webargparseargparsertc_input.warning(f"Error while writing to clipboard: {e}")
-            return False
-    async def start_clipboard(self):
-        if self.enable_clipboard in ["true", "out"]:
-            logger_webrtc_input.info("starting clipboard monitor")
-            self.clipboard_running = True
-            last_data = ""
-            while self.clipboard_running:
-                curr_data = self.read_clipboard()
-                if curr_data and curr_data != last_data:
-                    logger_webrtc_input.info(
-                        "sending clipboard content, length: %d" % len(curr_data)
-                    )
-                    self.on_clipboard_read(curr_data)
-                    last_data = curr_data
-                await asyncio.sleep(0.5)
-            logger_webrtc_input.info("clipboard monitor stopped")
-        else:
-            logger_webrtc_input.info("skipping outbound clipboard service.")
-    def stop_clipboard(self):
-        logger_webrtc_input.info("stopping clipboard monitor")
-        self.clipboard_running = False
-    async def start_cursor_monitor(self):
-        if not self.xdisplay.has_extension("XFIXES"):
-            if self.xdisplay.query_extension("XFIXES") is None:
-                logger_webrtc_input.error(
-                    "XFIXES extension not supported, cannot watch cursor changes"
-                )
-                return
-        xfixes_version = self.xdisplay.xfixes_query_version()
-        logger_webrtc_input.info(
-            "Found XFIXES version %s.%s"
-            % (
-                xfixes_version.major_version,
-                xfixes_version.minor_version,
-            )
-        )
-        logger_webrtc_input.info("starting cursor monitor")
-        self.cursor_cache = {}
-        self.cursors_running = True
-        screen = self.xdisplay.screen()
-        self.xdisplay.xfixes_select_cursor_input(
-            screen.root, xfixes.XFixesDisplayCursorNotifyMask
-        )
-        logger_webrtc_input.info("watching for cursor changes")
-        try:
-            image = self.xdisplay.xfixes_get_cursor_image(screen.root)
-            self.cursor_cache[image.cursor_serial] = self.cursor_to_msg(
-                image, self.cursor_scale, self.cursor_size
-            )
-            self.on_cursor_change(self.cursor_cache[image.cursor_serial])
-        except Exception as e:
-            logger_webrtc_input.warning("exception from fetching cursor image: %s" % e)
-        while self.cursors_running:
-            if self.xdisplay.pending_events() == 0:
-                await asyncio.sleep(0.1)
-                continue
-            event = self.xdisplay.next_event()
-            if (event.type, 0) == self.xdisplay.extension_event.DisplayCursorNotify:
-                cache_key = event.cursor_serial
-                if cache_key in self.cursor_cache:
-                    if self.cursor_debug:
-                        logger_webrtc_input.warning(
-                            "cursor changed to cached serial: {}".format(cache_key)
-                        )
-                else:
-                    try:
-                        cursor = self.xdisplay.xfixes_get_cursor_image(screen.root)
-                        self.cursor_cache[cache_key] = self.cursor_to_msg(
-                            cursor, self.cursor_scale, self.cursor_size
-                        )
-                        if self.cursor_debug:
-                            logger_webrtc_input.warning(
-                                "New cursor: position={},{}, size={}x{}, length={}, xyhot={},{}, cursor_serial={}".format(
-                                    cursor.x,
-                                    cursor.y,
-                                    cursor.width,
-                                    cursor.height,
-                                    len(cursor.cursor_image),
-                                    cursor.xhot,
-                                    cursor.yhot,
-                                    cursor.cursor_serial,
-                                )
-                            )
-                    except Exception as e:
-                        logger_webrtc_input.warning(
-                            "exception from fetching cursor image: %s" % e
-                        )
-                self.on_cursor_change(self.cursor_cache.get(cache_key))
-        logger_webrtc_input.info("cursor monitor stopped")
-    def stop_cursor_monitor(self):
-        logger_webrtc_input.info("stopping cursor monitor")
-        self.cursors_running = False
-    def cursor_to_msg(self, cursor, scale=1.0, cursor_size=-1):
-        if cursor_size > -1:
-            target_width = cursor_size
-            target_height = cursor_size
-            xhot_scaled = int(cursor_size / cursor.width * cursor.xhot)
-            yhot_scaled = int(cursor_size / cursor.height * cursor.yhot)
-        else:
-            target_width = int(cursor.width * scale)
-            target_height = int(cursor.height * scale)
-            xhot_scaled = int(cursor.xhot * scale)
-            yhot_scaled = int(cursor.yhot * scale)
-        png_data_b64 = base64.b64encode(
-            self.cursor_to_png(cursor, target_width, target_height)
-        )
-        override = None
-        if sum(cursor.cursor_image) == 0:
-            override = "none"
-        return {
-            "curdata": png_data_b64.decode(),
-            "handle": cursor.cursor_serial,
-            "override": override,
-            "hotspot": {
-                "x": xhot_scaled,
-                "y": yhot_scaled,
-            },
-        }
-    def cursor_to_png(self, cursor, resize_width, resize_height):
-        with io.BytesIO() as f:
-            s = [((i >> b) & 0xFF) for i in cursor.cursor_image for b in [16, 8, 0, 24]]
-            im = Image.frombytes("RGBA", (cursor.width, cursor.height), bytes(s), "raw")
-            if cursor.width != resize_width or cursor.height != resize_height:
-                im = im.resize((resize_width, resize_height))
-            im.save(f, "PNG")
-            data = f.getvalue()
-            if self.cursor_debug:
-                with open("/tmp/cursor_%d.png" % cursor.cursor_serial, "wb") as debugf:
-                    debugf.write(data)
-            return data
-    async def stop_js_server(self):
-        await self.__js_disconnect()
-    async def on_message(self, msg):
-        toks = msg.split(",")
-        if toks[0] == "pong":
-            if self.ping_start is None:
-                logger_webrtc_input.warning("received pong before ping")
-                return
-            roundtrip = time.time() - self.ping_start
-            latency = (roundtrip / 2) * 1000
-            latency = float("%.3f" % latency)
-            self.on_ping_response(latency)
-        elif toks[0] == "kd":
-            self.send_x11_keypress(int(toks[1]), down=True)
-        elif toks[0] == "ku":
-            self.send_x11_keypress(int(toks[1]), down=False)
-        elif toks[0] == "kr":
-            self.reset_keyboard()
-        elif toks[0] in ["m", "m2"]:
-            relative = False
-            if toks[0] == "m2":
-                relative = True
-            try:
-                x, y, button_mask, scroll_magnitude = [int(i) for i in toks[1:]]
-            except:
-                x, y, button_mask, scroll_magnitude = 0, 0, self.button_mask, 0
-                relative = False
-            try:
-                self.send_x11_mouse(x, y, button_mask, scroll_magnitude, relative)
-            except Exception as e:
-                logger_webrtc_input.warning("failed to set mouse cursor: {}".format(e))
-        elif toks[0] == "p":
-            visible = bool(int(toks[1]))
-            logger_webrtc_input.info("Setting pointer visibility to: %s" % str(visible))
-            self.on_mouse_pointer_visible(visible)
-        elif toks[0] == "vb":
-            bitrate = int(toks[1])
-            logger_webrtc_input.info("Setting video bitrate to: %d" % bitrate)
-            self.on_video_encoder_bit_rate(bitrate)
-        elif toks[0] == "ab":
-            bitrate = int(toks[1])
-            logger_webrtc_input.info("Setting audio bitrate to: %d" % bitrate)
-            self.on_audio_encoder_bit_rate(bitrate)
-        elif toks[0] == "js":
-            if toks[1] == "c":
-                js_num = int(toks[2])
-                name = base64.b64decode(toks[3]).decode()[:255]
-                num_axes = int(toks[4])
-                num_btns = int(toks[5])
-                self.__js_connect(js_num, name, num_btns, num_axes)
-            elif toks[1] == "d":
-                js_num = int(toks[2])
-                await self.__js_disconnect(js_num)
-            elif toks[1] == "b":
-                js_num = int(toks[2])
-                btn_num = int(toks[3])
-                btn_val = float(toks[4])
-                self.__js_emit_btn(js_num, btn_num, btn_val)
-            elif toks[1] == "a":
-                js_num = int(toks[2])
-                axis_num = int(toks[3])
-                axis_val = float(toks[4])
-                self.__js_emit_axis(js_num, axis_num, axis_val)
-            else:
-                logger_webrtc_input.warning("unhandled joystick command: %s" % toks[1])
-        elif toks[0] == "cr":
-            if self.enable_clipboard in ["true", "out"]:
-                data = self.read_clipboard()
-                if data:
-                    logger_webrtc_input.info(
-                        "read clipboard content, length: %d" % len(data)
-                    )
-                    self.on_clipboard_read(data)
-                else:
-                    logger_webrtc_input.warning("no clipboard content to send")
-            else:
-                logger_webrtc_input.warning(
-                    "rejecting clipboard read because outbound clipboard is disabled."
-                )
-        elif toks[0] == "cw":
-            if self.enable_clipboard in ["true", "in"]:
-                data = base64.b64decode(toks[1]).decode("utf-8")
-                self.write_clipboard(data)
-                logger_webrtc_input.info(
-                    "set clipboard content, length: %d" % len(data)
-                )
-            else:
-                logger_webrtc_input.warning(
-                    "rejecting clipboard write because inbound clipboard is disabled."
-                )
-        elif toks[0] == "r":
-            res = toks[1]
-            if re.match(re.compile(r"^\d+x\d+$"), res):
-                w, h = [int(i) + int(i) % 2 for i in res.split("x")]
-                self.on_resize("%dx%d" % (w, h))
-            else:
-                logger_webrtc_input.warning(
-                    "rejecting resolution change, invalid WxH resolution: %s" % res
-                )
-        elif toks[0] == "s":
-            scale = toks[1]
-            if re.match(re.compile(r"^\d+(\.\d+)?$"), scale):
-                self.on_scaling_ratio(float(scale))
-            else:
-                logger_webrtc_input.warning(
-                    "rejecting scaling change, invalid scale ratio: %s" % scale
-                )
-        elif toks[0] == "_arg_fps":
-            fps = int(toks[1])
-            logger_webrtc_input.info("Setting framerate to: %d" % fps)
-            self.on_set_fps(fps)
-        elif toks[0] == "_arg_resize":
-            if len(toks) != 3:
-                logger_webrtc_input.error(
-                    "invalid _arg_resize command, expected 2 arguments <enabled>,<resolution>"
-                )
-            else:
-                enabled = toks[1].lower() == "true"
-                logger_webrtc_input.info("Setting enable_resize to : %s" % str(enabled))
-                res = toks[2]
-                if re.match(re.compile(r"^\d+x\d+$"), res):
-                    w, h = [int(i) + int(i) % 2 for i in res.split("x")]
-                    enable_res = "%dx%d" % (w, h)
-                else:
-                    logger_webrtc_input.warning(
-                        "rejecting enable resize with resolution change to invalid resolution: %s"
-                        % res
-                    )
-                    enable_res = None
-                self.on_set_enable_resize(enabled, enable_res)
-        elif toks[0] == "_f":
-            try:
-                fps = int(toks[1])
-                self.on_client_fps(fps)
-            except:
-                logger_webrtc_input.error(
-                    "failed to parse fps from client: " + str(toks)
-                )
-        elif toks[0] == "_l":
-            try:
-                latency_ms = int(toks[1])
-                self.on_client_latency(latency_ms)
-            except:
-                logger_webrtc_input.error(
-                    "failed to parse latency report from client" + str(toks)
-                )
-        elif toks[0] == "_stats_video" or toks[0] == "_stats_audio":
-            try:
-                await self.on_client_webrtc_stats(toks[0], ",".join(toks[1:]))
-            except:
-                logger_webrtc_input.error(
-                    "failed to parse WebRTC Statistics JSON object"
-                )
-        elif toks[0] == "co":
-            if toks[1] == "end":
-                try:
-                    subprocess.run(["xdotool", "type", toks[2]], check=True)
-                except subprocess.CalledProcessError as e:
-                    logger_webrtc_input.warning(f"Error calling xdotool type: {e}")
-                except FileNotFoundError:
-                    logger_webrtc_input.warning(f"Xdotool not found on this system: {e}")
-        else:
-            logger_webrtc_input.info("unknown data channel message: %s" % msg)
 class WebRTCSignallingError(Exception):
     pass
 class WebRTCSignallingErrorNoPeer(Exception):
@@ -4057,6 +3249,11 @@ async def main():
         "--port",
         default=os.environ.get("SELKIES_PORT", "8080"),
         help='Port to listen to for the signaling and web server, default: "8080"',
+    )
+    parser.add_argument(
+        "--data_websocket_port",
+        default=os.environ.get("SELKIES_DATA_WEBSOCKET_PORT", "8082"),
+        help='Port to listen to for the raw data websocket server, default: "8082"',
     )
     parser.add_argument(
         "--web_root",
@@ -4487,6 +3684,8 @@ async def main():
     congestion_control = args.congestion_control.lower() == "true"
     video_packetloss_percent = float(args.video_packetloss_percent)
     audio_packetloss_percent = float(args.audio_packetloss_percent)
+
+    data_websocket_server = DataStreamingServer(int(args.data_websocket_port))
     event_loop = asyncio.get_running_loop()
     app = GSTWebRTCApp(
         event_loop,
@@ -4502,6 +3701,7 @@ async def main():
         congestion_control,
         video_packetloss_percent,
         audio_packetloss_percent,
+        data_streaming_server=data_websocket_server
     )
     audio_app = GSTWebRTCApp(
         event_loop,
@@ -4552,7 +3752,9 @@ async def main():
     signalling.on_session = on_session_handler
     audio_signalling.on_session = on_session_handler
     cursor_scale = 1.0
+    # Initialize WebRTCInput, passing the gst_webrtc_app instance
     webrtc_input = WebRTCInput(
+        app,
         args.uinput_mouse_socket,
         args.js_socket_path,
         args.enable_clipboard.lower(),
@@ -4724,6 +3926,7 @@ async def main():
     rtc_file_mon.on_rtc_config = mon_rtc_config
     try:
         asyncio.create_task(server.run())
+        asyncio.create_task(data_websocket_server.run_server())
         if using_metrics_http:
             asyncio.create_task(metrics.start_http())
         await webrtc_input.connect()
@@ -4763,6 +3966,7 @@ async def main():
         await rtc_file_mon.stop()
         system_mon.stop()
         await server.stop()
+        await data_websocket_server.stop()
         sys.exit(0)
 def entrypoint():
     asyncio.run(main())
